Add master volume control slider
Adds support for a master volume control to the HomeScreen app.
The master volume slider appears when touching in the MediaArea
at the bottom of the screen. If there is no interaction with the
slider, it fades out after 5 seconds so that the AGL logo area
is not obscured. The volume slider is accessible across all
applications due to the MediaArea portion of HomeScreen always
being exposed to the screen.

The slider volume control is tied to the default PA sink and
applies proportional volume control across all channels of the
default sink when changed.  It also reacts to external volume
change events from other PA clients (e.g. Mixer app or pactl
CLI changes) and updates the master volume slider accordingly.

AGL-Bug: SPEC-550
Change-Id: I9bf55bc624f4bb95c162a79bd2eb314d8f945033
Signed-off-by: Matt Porter <mporter@konsulko.com>