Fix panel backgrounds and sizes
The split into 3 windows done for the switch to the new compositor
dropped the full screen image background that had previously been
used, but did not set any replacement for the new top and bottom panel
windows.  This, combined with some rows of transparent pixels at the
bottom and top of the respective panel background images, resulted in
the default white background being visible in those areas.

To fix this, the background of the panel windows has been set to the
color used in the previous background image for those areas (#33363a).
Additionally:
- the top and bottom panel sizes have been rationalized to 216 pixels
  high each, since there is no longer any reason to have them be
  different sizes.
- the now unused main.qml has been removed to avoid confusion.

Bug-AGL: SPEC-3420

Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Change-Id: Ib656ff256f05362def378047171d1fe940355790