Add workaround for top panel button highlighting
The switch to the new compositor removed the callback to update the
top panel button highlight since it was being driven by the
Event_ScreenUpdated event from the old windowmanager.  For now, work
around this by driving the ApplicationLauncher's setCurrent method
from the appropriate place in the HomescreenHandler object's
tapShortcut method.  If a generic mechanism for notifications on
application expose becomes available via agl-shell-desktop, that
should be used instead.

Additionally, add an explicit call to setCurrent on initialization to
highlight the top panel Launcher button, matching the actual initial
UI state.

Bug-AGL: SPEC-3510

Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Change-Id: Ied5dd8e78195d061585510e60a758559ca4f69b3