main: Add different QMLs for doing screenshots
A better of way trying to verify that we have the homescreen
up-and-running would be to load-up a reference image which doesn't
change depending on time or other things loaded up (weather for
instance).

Alternatively, a possible solution would be to have a totally different
application that basically does the same thing as homescreen. Instead of
doing that, use an environment variable to choose up what QMLs to load
instead of the original application. Modifying the env variable then
re-starting the application is sufficient to have the screenshot for CI.

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>
Change-Id: Ie1478bb75f23c36e3c24c9e8d3e2d7c4848dd779