homescreenhandler: use applaunchd for app activation
As part of the App FW rework, we introduced a new app launcher service,
accessible through D-Bus. This service accepts requests to activate
applications, and notifies when a application successfully activated or
terminated.

Requests for activation are handled by a separate `launcher` application
for now, so we only need to listen on incoming signals in order to react
to those:
- when an application successfully started or is switched to, ask the
  compositor to activate it
- when an application terminates, activate the launcher so the user
  isn't facing a "blank" display

Bug-AGL: SPEC-4160

Signed-off-by: Arnaud Ferraris <arnaud.ferraris@collabora.com>
Change-Id: I7369944570651b45ec9dcca9ccde3dfd75719c3b