Use agl-shell-desktop protocol to keep track of closed windows
Currently, we have no way of knowing when the main window of an
application is closed unless this action terminates the app. Hooking up
the `agl-shell-desktop` protocol allows us to be notified when this
happens, so we can instruct `homescreenhandler` to switch back to the
previously active app.

Bug-AGL: SPEC-4222
Signed-off-by: Arnaud Ferraris <arnaud.ferraris@collabora.com>
Change-Id: I26dfaccce8894c2599afd9b14349b0703727d47a