homescreenhandler: Avoid passing an invalid appid
If there's no item added in the QList we'll return an invalid value
pointing to some invalid memory, which we can not determine if its empty
string or not, so this makes sure we check the list itself.

Found when looking at removal/activation of surfaces while re-doing some
parts in the compositor.

Bug-AGL: SPEC-4263

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>
Change-Id: I912aa5a42a7d21374a389a7f193979f30504a83b