homescreen: Add support for defining an activation area
This is an example on how to use it, and it enables a new variable
HOMESCREEN_EMBEDDED_PANELS to test it out. It gives out the same
rectangle are we have in the homescreen demo, but the panels are this
time embedded into the background surface, rather than being a separate
surface.

This introduces a new Qml file, background_with_panels which integrates
both the top and the bottom panel into a single qml, which main.cpp
loads.

While at it, this corrects up the indentation we have in shell.cpp file
and sets up the default output in homeScreenHandler constructor.

Due to this way integrates and works, it should *not* be visual change
between the two (the normal, two panel surfaces) and this version.

It can used as a example how to integrate this better, when managing
multiple surfaces is not easy to be done by the toolkit or runtime.

Bug-AGL: SPEC-4594
Signed-off-by: Marius Vlad <marius.vlad@collabora.com>
Change-Id: Ibe274b336b15e9713d0cdded5424cc8e92257d7f