homescreenhandler: Avoid using a stale wl_output
Rather than storing wl_output when creating a HomeScreenHandler object,
retrieve the output at activation. Presumably, we could use Screen and
from there get it but seems to be easier this way.

Re-plugging the connector back and forth would result in a stale wl_output,
so this would retrieve it always as the currently enabled/active one.

Bug-AGL: SPEC-4705
Signed-off-by: Marius Vlad <marius.vlad@collabora.com>
Change-Id: Id54c1f253b094e196c6bd367c7419c0ec5860d5f